+++ tests/input/test_slider.py
@@ -13,11 +13,12 @@ SLIDER_OPTIONS = {"minValue": 0, "maxValue": 100, "stepValue": 1, "selected": 50
 
 
 @pytest.fixture
-def dash_app() -> Callable[[Optional[int]], Dash]:
-    def app_factory(selected: Optional[int] = None) -> Dash:
+def dash_app() -> Callable[[Optional[int], Optional[str]], Dash]:
+    def app_factory(selected: Optional[int] = None, input_type: Optional[str] = None) -> Dash:
         kwargs: Dict[str, Any] = SLIDER_OPTIONS
         if selected is not None:
             kwargs["selected"] = selected
+        kwargs["inputType"] = input_type
 
         app = Dash(name=__name__)
         app.layout = mdc.Dashboard(
@@ -71,5 +72,57 @@ def test_slide_action(dash_duo: DashComposite, dash_app):
     slide_amount = random.randint(-50, 50)
     ActionChains(dash_duo.driver).drag_and_drop_by_offset(slider, slide_amount, 0).perform()
     assert dash_duo.find_element("#text").text == str(initial_value + slide_amount)
+
+    assert dash_duo.get_logs() is None
+
+
+def test_slider_render_float_input_text(dash_duo: DashComposite, dash_app):
+    dash_duo.start_server(dash_app(input_type="float", selected=12.34))
+    input_field = dash_duo.find_element("input")
+    assert input_field.get_attribute("value") == "12.34"
+    assert dash_duo.get_logs() is None
+
+
+def test_slider_render_integer_input_text(dash_duo: DashComposite, dash_app):
+    dash_duo.start_server(dash_app(input_type="float", selected=12))
+    input_field = dash_duo.find_element("input")
+    assert input_field.get_attribute("value") == "12"
+    assert dash_duo.get_logs() is None
+
+
+def test_slider_update_from_input_text(dash_duo: DashComposite, dash_app):
+    dash_duo.start_server(dash_app(input_type="float", selected=12.34))
+    input_field = dash_duo.find_element(".MuiInputBase-input")
+
+    # test if input text is reflected in output typography
+    input_field.clear()
+    input_field.send_keys("56.78")
+    assert dash_duo.find_element("#text").text == "56.78"
+
+    assert dash_duo.get_logs() is None
+
+
+def test_slider_input_validation(dash_duo: DashComposite, dash_app):
+    dash_duo.start_server(dash_app(input_type="float", selected=12.34))
+    input_field = dash_duo.find_element(".MuiInputBase-input")
+
+    # test if default precision (2) is respected
+    input_field.clear()
+    input_field.send_keys("12.345678")
+    assert dash_duo.find_element("#text").text == "12.34"
+
+    # test if invalid characters are ignored
+    input_field.clear()
+    input_field.send_keys("-")
+    input_field.send_keys(".")
+    assert dash_duo.find_element("#text").text == "12.34"
+    input_field.clear()
+    input_field.send_keys("0")
+    input_field.send_keys(".")
+    input_field.send_keys("abcd")
+    assert dash_duo.find_element("#text").text == "0"
+    input_field.clear()
+    input_field.send_keys("--")
+    assert dash_duo.find_element("#text").text == "0"
 
     assert dash_duo.get_logs() is None